Initial commit of APPROD
The software takes any deterministic simulation code (non-specific) and
modifies identified variables in the input-deck for the simulation code
before launching a "run" of the code. This process is repeated N times
and is useful for Probabilistic Risk Assessment studies as well as any
other study requiring many parametric changes to variables in a
deterministic code.